Create grover.py
