+++ src/qibo/models/__init__.py
@@ -1,4 +1,5 @@
 from qibo.models.circuit import Circuit, QFT
 from qibo.models.evolution import StateEvolution, AdiabaticEvolution
 from qibo.models.variational import VQE, QAOA
+from qibo.models.grover import Grover
 from qibo.models import hep
